+++ imaging/tools/custom/polylineScissorsTool.ts
